# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orthoplex API",
        "description": "A Laravel 12 multi-tenant SaaS API with comprehensive authentication, RBAC, and user management.",
        "contact": {
            "email": "support@orthoplex.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://orthoplex.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/webhooks": {
            "get": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "List webhooks",
                "description": "Get paginated list of webhooks with advanced filtering",
                "operationId": "2aa9e9009edaf4f064424d424b3bed7a",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "RSQL filter expression",
                        "schema": {
                            "type": "string",
                            "example": "active==true;events=in=user.created"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhooks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Webhook"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Create webhook",
                "description": "Create a new webhook endpoint",
                "operationId": "8bfcf5d6892778b0af84f9891a8cf892",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url",
                                    "events"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User Registration Hook"
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://api.example.com/webhooks"
                                    },
                                    "events": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "user.created",
                                            "user.updated"
                                        ]
                                    },
                                    "secret": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "headers": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "max_retries": {
                                        "type": "integer",
                                        "default": 3,
                                        "maximum": 10,
                                        "minimum": 0
                                    },
                                    "retry_delay": {
                                        "type": "integer",
                                        "default": 60,
                                        "maximum": 3600,
                                        "minimum": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Webhook created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Webhook"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/webhooks/{webhook}": {
            "get": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Get webhook",
                "description": "Get webhook details by ID",
                "operationId": "575c5890c17e0f76b43772046d04a6e7",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Webhook"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Update webhook",
                "description": "Update webhook configuration",
                "operationId": "61710d3e239969f292f2abfcc34f1f94",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "events": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "active": {
                                        "type": "boolean"
                                    },
                                    "headers": {
                                        "type": "object"
                                    },
                                    "max_retries": {
                                        "type": "integer"
                                    },
                                    "retry_delay": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Delete webhook",
                "description": "Delete a webhook endpoint",
                "operationId": "efe718f3bf61e800d3b188bdf6d1d536",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/webhooks/{webhook}/test": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Test webhook",
                "description": "Send a test payload to webhook endpoint",
                "operationId": "428d0308dc9f7a73e992ef35fdef6b08",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test webhook sent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/webhooks/events": {
            "get": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Get available events",
                "description": "Get list of available webhook events",
                "operationId": "4d6beddacc239a83d95602b86e825dd2",
                "responses": {
                    "200": {
                        "description": "Events retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/webhooks/{webhook}/regenerate-secret": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Regenerate webhook secret",
                "description": "Generate a new HMAC secret for webhook",
                "operationId": "15971d9a8b57045254dda69cbb916296",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secret regenerated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Webhook": {
                "title": "Webhook",
                "required": [
                    "id",
                    "name",
                    "url",
                    "events",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "User Registration Hook"
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://api.example.com/webhooks"
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "user.created",
                            "user.updated"
                        ]
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "secret": {
                        "type": "string",
                        "example": "abc123..."
                    },
                    "max_retries": {
                        "type": "integer",
                        "example": 3
                    },
                    "retry_delay": {
                        "type": "integer",
                        "example": 60
                    },
                    "last_triggered_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}